Added a patch to disable insertion of the word "dirty" into the version number.
This should "fix" a bug that causes the build to not run (due to
".dirty" getting appended to a release version number being invalid
according to semver spec). This is not the proper way to do it by
any means, but I don't have any sed experience and it doesn't look
like I'll be able to figure out how to use sed tonight.